Moving tokenize
